git-svn-id: http://svn.centreon.com/oreon-dev@2084 6bcd3966-0018-0410-8128-fd23d134de7e

+++ ODS/lib/verifyHostServiceIdName.pm
@@ -76,10 +76,14 @@ sub purgeRrdDB(){
     closedir DIR;
     for (@files) {
 		if (!defined($base{$_})){
-			if (unlink($some_dir."/".$_)){
-				writeLogFile("Warning : ".$some_dir."/".$_." removed \n");
-			} else {
-				writeLogFile("Error : Unable to remove ".$some_dir.$_ ."\n");
+			if (-d $some_dir."/".$_){
+				;
+			} else { 
+				if (unlink($some_dir."/".$_)){
+					writeLogFile("Warning : ".$some_dir."/".$_." removed \n");
+				} else {
+					writeLogFile("Error : Unable to remove ".$some_dir.$_ ."\n");
+				}
 			}
 		}
 	}
